requirements.txt
torch==2.4.1
torchaudio==2.4.1

transformers==4.44.2
accelerate==0.34.2
datasets==2.21.0
peft==0.12.0

jiwer==3.0.4
pyyaml==6.0.2
numpy==1.26.4

sounddevice==0.4.7

fastapi==0.112.2
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
